Move documentation-related modules under 'api_docs' subpackage
