+++ pyinaturalist/api_docs/__init__.py
@@ -1,0 +1,8 @@
+# flake8: noqa: F401, F403
+from pyinaturalist.api_docs.forge_utils import (
+    copy_doc_signature,
+    copy_docstrings,
+    copy_signature,
+    copy_signatures,
+    document_request_params,
+)
