+++ pyinaturalist/api_docs/templates.py
