+++ pyinaturalist/v0/observations.py
@@ -2,7 +2,8 @@
 from logging import getLogger
 from typing import List, Union
 
-from pyinaturalist import api_docs as docs
+from pyinaturalist.api_docs import document_request_params
+from pyinaturalist.api_docs import templates as docs
 from pyinaturalist.api_requests import delete, get, post, put
 from pyinaturalist.constants import (
     API_V0_BASE_URL,
@@ -13,7 +14,6 @@ from pyinaturalist.constants import (
 )
 from pyinaturalist.converters import convert_all_coordinates, convert_all_timestamps
 from pyinaturalist.exceptions import ObservationNotFound
-from pyinaturalist.forge_utils import document_request_params
 from pyinaturalist.pagination import add_paginate_all
 from pyinaturalist.request_params import (
     convert_observation_fields,
